notesdjango/requirements.txt
asgiref==3.8.1
attrs==23.2.0
axios==0.4.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
h11==0.14.0
idna==3.6
lxml==5.2.2
markdown-it-py==3.0.0
mdurl==0.1.2
outcome==1.3.0.post0
packaging==23.2
pycparser==2.21
Pygments==2.18.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
rich==13.7.1
selenium==4.17.2
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.5.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
webdriver-manager==4.0.1
wsproto==1.2.0
